requirements.txt
autogen-agentchat==0.2.0
aio-pika==9.3.0
fastapi==0.112.0
uvicorn[standard]==0.30.0
openai==1.12.0
posthog==3.4.0
PyGithub==2.1.0
loguru==0.7.2
chromadb==0.5.0
pydantic==2.6.0
python-dotenv==1.0.0
httpx==0.27.0
scipy==1.13.0
numpy==2.0.0
pydantic-yaml==0.6.0
